Move tests where they should belong
